Change sleep time to long, to match Poco methods.

+++ libs/openFrameworks/utils/ofThread.cpp
@@ -140,7 +140,7 @@ void ofThread::waitForThread(bool stop, long waitMS){
 }
 
 //-------------------------------------------------
-void ofThread::sleep(int sleepMS){
+void ofThread::sleep(long sleepMS){
 	Poco::Thread::sleep(sleepMS);
 }
